+++ CHANGELOG.md
@@ -5,11 +5,30 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+[1.0.0]: https://github.com/johnarp/proficiency-tracker/compare/v0.3.0...v1.0.0
 [0.3.0]: https://github.com/johnarp/proficiency-tracker/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/johnarp/proficiency-tracker/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/johnarp/proficiency-tracker/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/johnarp/proficiency-tracker/compare/v0.0.1...v0.1.0
 [0.0.1]: https://github.com/johnarp/proficiency-tracker/releases/tag/v0.0.1
+
+## [1.0.0] - 2026-03-18
+
+### Added
+
+- Web Warriors, Star Turn, Plans Within Plans, Regenesis, and Inner Devil Gallery Card Themes
+- Show proficiency level on card
+- Announcement, profile, and credit functionality
+- Reusable modals for announcements, profile, and credits
+
+### Changed
+
+- Hero name scales with card size
+- CSS improvements
+
+### Fixed
+
+- Hero level is now default to 1 instead of null
 
 ## [0.3.0] - 2026-03-15